backend/requirements.txt
amqp==2.6.0
asgiref==3.2.9
billiard==3.6.3.0
celery==4.4.6
Django==3.0.7
django-cors-headers==3.4.0
django-extensions==2.2.9
django-tastypie==0.14.3
dnspython==1.16.0
email-validator==1.1.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2==2.8.5
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
